backend/pyproject.toml
[tool.poetry]
name = "deepchatai-backend"
version = "0.1.0"
description = "AI-powered chat application backend"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
tortoise-orm = {extras = ["asyncpg", "aiosqlite"], version = "^0.20.0"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.6"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
envparse = "^0.2.0"
langchain = "^0.1.0"
langchain-openai = "^0.0.5"
langchain-community = "^0.0.13"
langchain-core = "^0.1.15"
openai = "^1.6.1"
faiss-cpu = "^1.7.4"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.11.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
